backend/requirements.txt
fastapi
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.4
opencv-python==4.10.1.26
scipy==1.13.1
ultralytics==8.0.226
pydicom==2.4.4
# PyTorch with pre-built wheels (no ninja compilation needed)
torch==2.1.1+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.1+cpu --index-url https://download.pytorch.org/whl/cpu
python-dotenv==1.0.0
